refactor(BE): TrackRegister 패키지 구조 수정 TrackRegister 패키지를 track/service 에서 공통 관리 및 클래스 이름 변경
