+++ backend/src/main/java/dev/be/dorothy/track/service/TrackRegisterService.java
@@ -1,0 +1,7 @@
+package dev.be.dorothy.track.service;
+
+import dev.be.dorothy.member.MemberRole;
+
+public interface TrackRegisterService {
+    TrackResDto create(Long memberIdx, String name, MemberRole role);
+}
