--- backend/src/main/java/dev/be/dorothy/track/service/register/TrackRegister.java
@@ -1,8 +1,0 @@
-package dev.be.dorothy.track.service.register;
-
-import dev.be.dorothy.member.MemberRole;
-import dev.be.dorothy.track.service.TrackResDto;
-
-public interface TrackRegister {
-    TrackResDto create(Long memberIdx, String name, MemberRole role);
-}
